libcameraservice: support disabling torch control support
On some devices such as OnePlus 12, newer blobs have a strange crash
which is entirely proprietary stacktraced with torch control support,
and Oplus stock disables in platform.

We previously had a hack to manually implement it in an extension,
but this device was also the only one that had other issues with it.

Additionally, >2.3 aidl isn't expected to lack this support. So this
hack simply always uses the same strength (legacy path), and does
not use torch control.

Change-Id: If0679f4b955b997c7afc4440729c6b84bbed1867
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>
Signed-off-by: Jackeagle <jackeagle102@gmail.com>